+++ printing/android/src/main/java/net/nfet/flutter/printing/PrintingJob.java
@@ -73,13 +73,16 @@ public class PrintingJob extends PrintDocumentAdapter {
     }
 
     static HashMap<String, Object> printingInfo() {
+        final boolean canPrint = android.os.Build.VERSION.SDK_INT >= Build.VERSION_CODES.KITKAT;
+        final boolean canRaster = Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP;
+
         HashMap<String, Object> result = new HashMap<>();
         result.put("directPrint", false);
-        result.put("dynamicLayout", true);
-        result.put("canPrint", true);
-        result.put("canConvertHtml", true);
+        result.put("dynamicLayout", canPrint);
+        result.put("canPrint", canPrint);
+        result.put("canConvertHtml", canRaster);
         result.put("canShare", true);
-        result.put("canRaster", true);
+        result.put("canRaster", canRaster);
         return result;
     }
 
@@ -181,8 +184,7 @@ public class PrintingJob extends PrintDocumentAdapter {
         thread.start();
     }
 
-    void printPdf(@NonNull String name, Double width, Double height, Double marginLeft,
-            Double marginTop, Double marginRight, Double marginBottom) {
+    void printPdf(@NonNull String name) {
         jobName = name;
         printJob = printManager.print(name, this, null);
     }